+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+ * Update minimum supported Rust version to 1.97.0
+
 ## [v1.0.1] - 2026-07-10
 
  * Fix wrong "install time ago" output when a package's install date is in the